requirements.txt
pandas
evidently==0.4.13
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0
imbalanced-learn>=0.8
matplotlib>=3.4
seaborn>=0.11
statsmodels>=0.12
mlflow
dagshub
streamlit
watchdog
imblearn
XGBOOST
python-dotenv
scipy